Added additional question, startpage moved to higher level
